More reliance on parser result
- Moved print pocket extra info to helpers partial
- Moved pause statements off Image Pack Selector to right after the menu call
- Removed the With Parsed and dependencies on local variables
- Leveraged the common Option path to reduce local variables
- Used global helper more after initialize
